transcript-downloader/requirements.txt
flask
playwright
pyyaml
